+++ src/products/store/actions/toppings.action.spec.ts
@@ -4,7 +4,7 @@ describe('Toppings Actions', () => {
   describe('LoadToppings Actions', () => {
     describe('LoadToppings', () => {
       it('should create an action', () => {
-        const action = fromHandlers.LoadToppings.action();
+        const action = fromHandlers.LoadToppings.create();
         expect({ ...action }).toEqual({
           type: fromHandlers.LoadToppings.type,
         });
@@ -14,7 +14,7 @@ describe('Toppings Actions', () => {
     describe('LoadToppingsFail', () => {
       it('should create an action', () => {
         const payload = { message: 'Load Error' };
-        const action = fromHandlers.LoadToppingsFail.action(payload);
+        const action = fromHandlers.LoadToppingsFail.create(payload);
 
         expect({ ...action }).toEqual({
           type: fromHandlers.LoadToppingsFail.type,
@@ -30,7 +30,7 @@ describe('Toppings Actions', () => {
           { id: 2, name: 'mushroom' },
           { id: 3, name: 'basil' },
         ];
-        const action = fromHandlers.LoadToppingsSuccess.action(payload);
+        const action = fromHandlers.LoadToppingsSuccess.create(payload);
 
         expect({ ...action }).toEqual({
           type: fromHandlers.LoadToppingsSuccess.type,
@@ -43,7 +43,7 @@ describe('Toppings Actions', () => {
   describe('VisualiseToppings Actions', () => {
     describe('VisualiseToppings', () => {
       it('should create an action', () => {
-        const action = fromHandlers.VisualiseToppings.action([1, 2, 3]);
+        const action = fromHandlers.VisualiseToppings.create([1, 2, 3]);
         expect({ ...action }).toEqual({
           type: fromHandlers.VisualiseToppings.type,
           payload: [1, 2, 3],
