+++ tsconfig.json
@@ -17,11 +17,6 @@
     "rootDir": ".",
     "sourceMap": true,
     "skipLibCheck": true,
-    "target": "es5",
-    "paths": {
-      "@ubik/*": [
-        "ubik_modules/*"
-      ]
-    }
+    "target": "es5"
   }
 }
